Move GoalView styles
